backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pandas>=2.3.2
openai>=1.40.0
python-multipart>=0.0.9
pyyaml>=6.0.1
